requirements.txt
streamlit==1.28.1
Pillow>=10.0.0
pandas==2.1.3
numpy==1.24.3
bcrypt>=4.0.0

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Analytics
schedule==1.2.0
plotly>=5.17.0

# Notion Integration
notion-client==2.2.1

# OpenAI
openai>=1.0.0

# Environment Variables
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Web Scraping & SEO Analysis
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# JavaScript Execution (Optional - for dynamic structured data)
# Uncomment the following lines if you want to use Selenium or Playwright
# selenium>=4.15.0
playwright>=1.40.0